server/requirements.txt
attrs==22.2.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
google-ads==19.0.0
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-secret-manager==2.15.0
google-crc32c==1.5.0
googleads==36.0.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.48.2
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
oauthlib==3.2.2
platformdirs==2.6.2
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
-e git+https://github.com/tahsin52/GoogleAdsOauth.git@64474feb9d44ed75c7c8b2596895405ea45b0f83#egg=UNKNOWN
urllib3==1.26.14
Werkzeug==2.2.2
xmltodict==0.13.0
zeep==4.2.1
zipp==3.11.0
